+++ R/TreeDist-package.R
@@ -1,14 +1,21 @@
 #' TreeDist:  Distances between Phylogenetic Trees
 #' 
-#' 'TreeDist' implements a suite of metrics that quantify the topological 
-#' distance between pairs of unweighted phylogenetic trees.
-#' The metrics generally fall in the category of 'generalized Robinson-Foulds
-#' distances': they are based on comparing partitions between trees, and thus
-#' reflect the relationship data within trees, with no reference to branch lengths.
+#' 'TreeDist' is an R package that implements a suite of metrics that quantify the
+#' topological distance between pairs of unweighted phylogenetic trees.
+#' It also includes a simple 'Shiny' application to allow the visualization of
+#' distance-based tree spaces.
+#' 
+#' 'TreeDist' primarily employs metrics in the category of
+#' 'generalized Robinson\ifelse{html}{&#ndash;}{--}Foulds distances': they are based on comparing splits
+#' (bipartitions) between trees, and thus reflect the relationship data within 
+#' trees, with no reference to branch lengths.
 #' Detailed documentation and usage instructions are 
 #' [available online](https://ms609.github.io/TreeDist/) or in the vignettes.
 #' 
-#' The [Robinson-Foulds distance](https://ms609.github.io/TreeDist/articles/Robinson-Foulds.html)
+#' 
+#' ## Generalized RF distances
+#' 
+#' The [Robinson\ifelse{html}{&#ndash;}{--}Foulds distance](https://ms609.github.io/TreeDist/articles/Robinson-Foulds.html)
 #' simply tallies the number of non-trivial splits (sometimes inaccurately
 #' termed clades, nodes or edges) that occur in both trees -- any splits that are
 #' not perfectly identical contributes one point to the distance score of zero, 
@@ -41,7 +48,7 @@
 #'   + Nye _et al._ (2006) score matchings according to the size of the largest 
 #'     split that is consistent with both of them, normalized against 
 #'     the Jaccard index.  This approach is extended by B&ouml;cker _et al_. (2013)
-#'     with the Jaccard-Robinson-Foulds metric (function 
+#'     with the Jaccard\ifelse{html}{&#ndash;}{--}Robinson\ifelse{html}{&#ndash;}{--}Foulds metric (function 
 #'     [`JaccardRobinsonFoulds()`](https://ms609.github.io/TreeDist/reference/JaccardRobinsonFoulds.html)).
 #' 
 #' * [`MatchingSplitDistance()`](https://ms609.github.io/TreeDist/reference/MatchingSplitDistance.html)
@@ -67,20 +74,34 @@
 #' the package '[TBRDist](https://ms609.github.io/TBRDist/index.html)'.
 #' 
 #' 
+#' # Tree space analysis
+#' 
+#' Project tree spaces and readily visualize projected landscapes, avoiding
+#' common analytical pitfalls (Smith, forthcoming),
+#' using the inbuilt graphical user interface:
+#'   
+#'   ```r
+#' TreeDist::Project()
+#' ```
+#' 
+#' Serious analysts should consult the
+#' [vignette](https://ms609.github.io/TreeDist/articles/treespace.html)
+#' for a command-line interface.
+#' 
 #' 
 #' @seealso
 #' 
 #' Further documentation is available in the 
 #' [package vignettes](https://ms609.github.io/TreeDist/articles/), visible from
-#' R using `vignette(package='TreeDist')`.
+#' R using `vignette(package = 'TreeDist')`.
 #' 
 #' Other R packages implementing tree distance functions include:
 #'   
 #'   * [ape](http://ape-package.ird.fr/):
 #'     - `cophenetic.phylo()`: Cophenetic distance
-#'     - `dist.topo()`: Path (topological) distance, Robinson-Foulds distance.
+#'     - `dist.topo()`: Path (topological) distance, Robinson\ifelse{html}{&#ndash;}{--}Foulds distance.
 #'   * [phangorn](https://cran.r-project.org/package=phangorn)
-#'     - `treedist()`: Path, Robinson-Foulds and approximate SPR distances.
+#'     - `treedist()`: Path, Robinson\ifelse{html}{&#ndash;}{--}Foulds and approximate SPR distances.
 #'   * [Quartet](https://ms609.github.io/Quartet/): Triplet and Quartet distances, 
 #'   using the tqDist algorithm.
 #'   * [TBRDist](https://ms609.github.io/TBRDist/): TBR and SPR distances on 
@@ -104,6 +125,9 @@
 #' 
 #' - \insertRef{SmithDist}{TreeDist}
 #' 
+#' - \insertRef{SmithSpace}{TreeDist}
+#' 
+#' @encoding UTF-8
 #' @keywords internal
 "_PACKAGE"